Make signed/unsigned conversions in the grounder explicit
clingo's own code under lib/ and app/ had a scattering of implicit
signed<->unsigned conversions: iterator distances and std::distance results
used directly as sizes or indices, size_t counts advanced through signed
iterator arithmetic, and a handful of int/unsigned mismatches in comparisons.
Each instance compiles just fine, but it makes things difficult to both
understand and reason about since each conversion requires a kind of mental
reset, and it feels error prone in the long term.

This brings lib/ and app/ to a clean build under -Wconversion and
-Wsign-conversion by fixing each site the right way rather than blanket-casting:

- Use unsigned index/size types, and std::cmp_* for the mixed-sign comparisons,
  where that is the correct fix.
- At the boundaries that genuinely cross signedness (a distance used as a size,
  a literal known positive by an assert, a freshly minted id), make the
  conversion explicit with a static_cast and a one-line comment naming the
  invariant that makes it safe.
- Guard the re2c-generated lexer include with a diagnostic pragma; the generated
  scanner reads a char buffer as unsigned char and is not ours to change, but
  the hand-written code in that translation unit stays covered.

No behavior change: every conversion here was already happening implicitly and
in range; the patch only makes the intent visible to the compiler and the
reader. Two atom-id/literal casts are left with a FIXME because that pair is a
genuine type conflation better solved by distinct atom and literal types than by
a cast; they are called out rather than papered over. All tests still pass.

+++ lib/control/include/clingo/control/solver.hh
@@ -194,7 +194,7 @@ class TermBaseMap {
     //!
     //! @param sym the symbol
     //! @return the index
-    [[nodiscard]] auto index(Symbol sym) const -> size_t { return map_.find(sym) - map_.begin(); }
+    [[nodiscard]] auto index(Symbol sym) const -> size_t { return static_cast<size_t>(map_.find(sym) - map_.begin()); }
 
     //! Get the number of mapped symbols.
     //!
